Move Edit Mode tests for Attributes
